set the endpoints in users-rotuers and auth-routers using express routing
