+++ auth/auth-router.js
@@ -1,0 +1,38 @@
+const router = require("express").Router();
+const bcrypt = require("bcryptjs");
+
+const UsersDb = require("../users/users-model.js");
+
+router.post("/register", (req, res) => {
+  let user = req.body;
+
+  const hash = bcrypt.hashSync(user.password, 8);
+  user.password = hash;
+
+  UsersDb.add(user)
+    .then(saved => {
+      res.status(201).json(saved);
+    })
+    .catch(error => {
+      res.status(500).json(error);
+    });
+});
+
+router.post("/login", (req, res) => {
+  let { username, password } = req.body;
+
+  UsersDb.findBy({ username })
+    .first()
+    .then(user => {
+      if (user && bcrypt.compareSync(password, user.password)) {
+        res.status(200).json({ message: `Welcome ${user.username}!` });
+      } else {
+        res.status(401).json({ message: "Invalid Credentials" });
+      }
+    })
+    .catch(error => {
+      res.status(500).json(error);
+    });
+});
+
+module.exports = router;
